Use old kyma cli because newest doesn't support minikube. Disable removing image when testing provisioning because image is not created then. Only vm instance is provisioned.

+++ development/custom-image/create-custom-image.sh
@@ -13,7 +13,7 @@ source "${ROOT_DIR}/prow/scripts/lib/gcp.sh"
 cleanup() {   
     log::info "Removing instance $VM_NAME"
     gcloud compute instances delete --quiet --zone "${ZONE}" "$VM_NAME"
-    if [ "$JOB_TYPE" == "presubmit" ]; then
+    if [[ "$JOB_TYPE" == "presubmit" && "$testMinikube" != "true" && "$testK3d" != "true" ]]; then
       log::info "Removing image $IMAGE"
       gcloud compute images delete "$IMAGE"
     fi
@@ -102,8 +102,9 @@ utils::ssh_to_vm_with_script -z "${ZONE}" -n "${VM_NAME}" -c "sudo sh -c 'mv /tm
 if [[ $testMinikube == true ]]; then
     log::info "Testing minikube installation"
     log::info "Download stable Kyma CLI"
-    utils::ssh_to_vm_with_script -z "${ZONE}" -n "${VM_NAME}" -c "curl -Lo kyma https://storage.googleapis.com/kyma-cli-stable/kyma-linux"
-    utils::ssh_to_vm_with_script -z "${ZONE}" -n "${VM_NAME}" -c "chmod +x kyma && mkdir ./bin && mv ./kyma ./bin/kyma && sudo cp ./bin/kyma /usr/local/bin/kyma"
+    # Using old version of kyma cli because newest doesn't support provision on minikube.
+    utils::ssh_to_vm_with_script -z "${ZONE}" -n "${VM_NAME}" -c "curl -sSLo kyma.tar.gz \"https://github.com/kyma-project/cli/releases/download/1.24.8/kyma_linux_x86_64.tar.gz\""
+    utils::ssh_to_vm_with_script -z "${ZONE}" -n "${VM_NAME}" -c "tar xvzf kyma.tar.gz && chmod +x kyma && mkdir ./bin && mv ./kyma ./bin/kyma && sudo cp ./bin/kyma /usr/local/bin/kyma"
     log::info "Triggering Minikube installation"
     utils::ssh_to_vm_with_script -z "${ZONE}" -n "${VM_NAME}" -c "yes | kyma provision minikube"
 fi
